+++ src/AppConstants.ts
