+++ src/RangeSelector.ts
