+++ src/app.js
@@ -4,7 +4,7 @@
 import * as plugins from 'phovea_core/src/plugin';
 import * as d3 from 'd3';
 import * as events from 'phovea_core/src/event';
-import { AppConstants } from './app_constants';
+import { AppConstants } from './AppConstants';
 /**
  * The main class for the app
  */
